refactor signal_handlers
