+++ server/signal-handlers.h
@@ -1,0 +1,6 @@
+#include "server/php-runner.h"
+
+
+void perform_error_if_running(const char *msg, script_error_t error_type);
+
+void init_handlers();
